services/gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
opensearch-py==2.4.0
qdrant-client==1.7.0
psycopg2-binary==2.9.9
minio==7.2.0
confluent-kafka==2.3.0
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.2
python-dotenv==1.0.0
prometheus-client==0.19.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
httpx==0.25.2
